Update cache and tests
